+++ packages/components-js/projects/jsdom-polyfill/tests/unit/specs/segmented-control.spec.ts
@@ -6,7 +6,7 @@ it('should have initialized shadow dom', async () => {
   document.body.innerHTML = getMarkup('p-segmented-control');
   expect(await componentsReady()).toBe(3);
 
-  const els = document.body.querySelectorAll('*');
+  const els = document.body.querySelectorAll('p-segmented-control,p-segmented-control-item');
   expect(els.length).toBe(3);
   els.forEach((el) => {
     expect(el.shadowRoot).not.toBeNull();